Allow editing of shortcut items.
Editing of a shortcut is initiated by holding a touch on the tile for a moment (this isn’t very discoverable, but neither is the Springboard’s editing…).
A menu will pop up allowing the user to remove the shortcut, or set it to show the main menu or keyboard, or execute a custom key sequence.

Currently there is no interface to adding new shortcuts. I think this will be done by having one empty tile always visible at the end of the list (on a page of its own if the last page is full), which the user can hold to edit in the normal way.
Alternatively we could have a new option in the edit menu, “Insert New Shortcut”. This might be preferable as then you could insert shortcuts in the first page if you wanted, but that menu is already getting a litle big.